+++ CHANGELOG.md
@@ -6,13 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [7.2.5] - 2026-09-TBD
 
+### Added
+
+- OSV Scanning support.
+- Actor auditing (checks if they're dead).
+
 ### Changed
 
 - Improved show-audit tools.
 
-### Added
+### Fixed
 
-- OSV Scanning support
+- TMDB scores now work properly.
 
 ## [7.2.4] - 2026-09-14